Feat/auth user socialing 유저 엔티티 dongCode 타입 변경

+++ src/main/java/com/example/backend/dto/user/SnsUserDTO.java
@@ -21,7 +21,7 @@ public class SnsUserDTO {
         this.email = email;
     }
 
-    public User toEntity(UserType usertype, String username, String phoneNumber, int dongCode, String dongName) {
+    public User toEntity(UserType usertype, String username, String phoneNumber, Long dongCode, String dongName) {
         return User.builder()
                 .userType(usertype).email(email).nickname(nickname)
                 .username(username).phoneNumber(phoneNumber)
