+++ src/main/java/com/example/backend/dto/user/UserDefaultJoinRequestDTO.java
@@ -42,7 +42,7 @@ public class UserDefaultJoinRequestDTO {
 
     @Schema(description = "지역 번호 8자리", defaultValue = "1111010400")
     @NotNull
-    private Integer dongCode;
+    private Long dongCode;
 
     @Schema(description = "행정구역명", defaultValue = "서울특별시 종로구 효자동")
     @NotNull
